modules/analysis/rosbag-image-pipeline/requirements-dev.txt
# You will need to verify the proper versions
# of the libraries for your use case.
# Check the version of MWAA you are using and compare
# to these versions

pytest-spark
apache-airflow~=2.5.2
apache-airflow-providers-amazon~=7.2.1
mypy-boto3-sqs
mypy-boto3-batch
